Support different way of adding member functions
Different version of CW did the right thing this time where they add
member functions as class/struct members, rather than making them global
and not even adding access level attributes to it.

Though when I say "the right thing" take this with a heavy grain of
salt, as apparently it was at the expense of parameter names. Though
perhaps this game was compiled with different debug parameters?